'|' and '_'  s around the map
just a quick change so that there are ! and _ s around the map so we can easily detect if it is the edge of the map. I tried to enable an ‾ at the bottom layer so that it looks smoother, but no matter how I formatted it either java didn't run it or the shell didn't show it. It is not really that important so we can always come back and mess with it if we have time.

Now that I think about it we can probably just check if it is the first or last element of the array but we can always change it back and this is easier

also changed move() to movePlayer() in main, as I need to make a move() for the enemies. We might want to refactor so that move() can be used for everything.