added killEnemy() method
Note, you need to find out which enemy it is which is dying. I'm not sure how one might do this Osamah, but if you can't figure it out lmk and I'll come up with something, probably an enemyAt class which will just check to see if there is an enemy at that spot and if so which.